Move DSL::Colour, DSL::Style and Alignment into dsl/shared.
