+++ lib/vedeu/dsl/all.rb
@@ -1,6 +1,5 @@
-require 'vedeu/dsl/shared'
-require 'vedeu/dsl/colour'
-require 'vedeu/dsl/style'
+require 'vedeu/dsl/shared/all'
+
 require 'vedeu/dsl/border'
 require 'vedeu/dsl/composition'
 require 'vedeu/dsl/geometry'
