+++ lib/vedeu/dsl/shared/all.rb
@@ -1,0 +1,6 @@
+require 'vedeu/dsl/shared/alignment'
+require 'vedeu/dsl/shared/colour'
+require 'vedeu/dsl/shared/style'
+
+module Vedeu
+end # Vedeu
