+++ lib/vedeu/dsl/shared/colour.rb
