+++ lib/vedeu/dsl/shared/style.rb
