+++ test/lib/vedeu/dsl/shared/alignment_test.rb
@@ -1,0 +1,26 @@
+require 'test_helper'
+
+module Vedeu
+
+  module DSL
+
+    describe Alignment do
+
+      # describe '#align' do
+      #   let(:value)   { 'silicon' }
+      #   let(:options) { {} }
+
+      #   subject { described.align(value, options) }
+
+      #   it { subject.must_be_instance_of(String) }
+
+      #   it 'aligns the value according to the options provided' do
+      #     subject.must_equal('siliconsilicon')
+      #   end
+      # end
+
+    end # Alignment
+
+  end # DSL
+
+end # Vedeu
