+++ test/lib/vedeu/dsl/shared/colour_test.rb
