+++ test/lib/vedeu/dsl/shared/style_test.rb
