--- test/lib/vedeu/support/alignment_test.rb
@@ -1,22 +1,0 @@
-require 'test_helper'
-
-module Vedeu
-
-  describe Alignment do
-
-    # describe '#align' do
-    #   let(:value)   { 'silicon' }
-    #   let(:options) { {} }
-
-    #   subject { described.align(value, options) }
-
-    #   it { subject.must_be_instance_of(String) }
-
-    #   it 'aligns the value according to the options provided' do
-    #     subject.must_equal('siliconsilicon')
-    #   end
-    # end
-
-  end # Alignment
-
-end # Vedeu
